add test and spy for webView
