+++ ImageFeedTests/Helpers/ImagesListPresenterSpy.swift
