+++ ImageFeedTests/WebViewPresenterTests.swift
@@ -1,0 +1,35 @@
+import XCTest
+@testable import ImageFeed
+
+final class WebViewTests: XCTestCase {
+    
+    func testViewControllerCallsLoadAuthView() {
+        // given
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        let viewController = storyboard.instantiateViewController(withIdentifier: "WebViewViewController") as! WebViewViewController
+        let presenter = WebViewPresenterSpy()
+        viewController.presenter = presenter
+        presenter.view = viewController
+        
+        // when
+        _ = viewController.view
+        
+        // then
+        XCTAssertTrue(presenter.loadAuthViewCalled)
+    }
+    
+    func testPresenterCallsLoadRequest() {
+        // given
+        let viewController = WebViewViewControllerSpy()
+        let authHelper = AuthHelper()
+        let presenter = WebViewPresenter(authHelper: authHelper)
+        viewController.presenter = presenter
+        presenter.view = viewController
+        
+        //when
+        presenter.loadAuthView()
+        
+        //then
+        XCTAssertTrue(viewController.loadRequestCalled)
+    }
+}
