--- ImageFeedTests/WebViewTests.swift
@@ -1,20 +1,0 @@
-import XCTest
-@testable import ImageFeed
-
-final class WebViewTests: XCTestCase {
-    
-    func testViewControllerCallsLoadAuthView() {
-        // given
-        let viewController = WebViewViewController() 
-        let presenter = WebViewPresenterSpy()
-        viewController.presenter = presenter
-        presenter.view = viewController
-        
-        // when
-        _ = viewController.view
-        
-        // then
-        XCTAssertTrue(presenter.loadAuthViewCalled)
-    }
-    
-}
